@hovey significant speedups to smoothing and metrics calculations by switching nodal coordinates to use `flavio`.

## Before

<pre>test block_16::calculate_laplacian                 ... <font color="#58C7E2">bench</font>:     146,297.06 ns/iter (+/- 3,967.91)
test block_16::smooth_laplace                      ... <font color="#58C7E2">bench</font>:     156,339.88 ns/iter (+/- 13,520.23)
test block_32::calculate_laplacian                 ... <font color="#58C7E2">bench</font>:     953,140.00 ns/iter (+/- 26,602.83)
test block_32::smooth_laplace                      ... <font color="#58C7E2">bench</font>:   1,166,944.30 ns/iter (+/- 43,341.31)
test block_8::calculate_laplacian                  ... <font color="#58C7E2">bench</font>:      22,111.04 ns/iter (+/- 2,365.38)
test block_8::smooth_laplace                       ... <font color="#58C7E2">bench</font>:      24,735.09 ns/iter (+/- 1,118.65)</pre>

<pre>test block_16::write_metrics_csv                   ... <font color="#58C7E2">bench</font>:   1,242,885.81 ns/iter (+/- 56,324.67)
test block_16::write_metrics_npy                   ... <font color="#58C7E2">bench</font>:   1,035,549.75 ns/iter (+/- 7,419.67)
test block_32::write_metrics_csv                   ... <font color="#58C7E2">bench</font>:   9,467,301.30 ns/iter (+/- 159,873.24)
test block_32::write_metrics_npy                   ... <font color="#58C7E2">bench</font>:   7,711,699.40 ns/iter (+/- 130,444.58)
test block_8::write_metrics_csv                    ... <font color="#58C7E2">bench</font>:     206,445.47 ns/iter (+/- 2,654.48)
test block_8::write_metrics_npy                    ... <font color="#58C7E2">bench</font>:     182,171.70 ns/iter (+/- 1,472.42)</pre>

<pre><font color="#005FD7">cargo</font> <font color="#00AFFF">run</font> <font color="#00AFFF">-qr</font> <font color="#00AFFF">-F</font> <font color="#00AFFF">profile</font> <font color="#00AFFF">--</font> <font color="#00AFFF">mesh</font> <font color="#00AFFF">-i</font> <font color="#00AFFF"><u style="text-decoration-style:single">spheres_resolution_3.npy</u></font> <font color="#00AFFF">-o</font> <font color="#00AFFF"><u style="text-decoration-style:single">target/foo.exo</u></font> <font color="#00AFFF">--metrics</font> <font color="#00AFFF"><u style="text-decoration-style:single">target/foo.npy</u></font> <font color="#00AFFF">smooth</font> <font color="#00AFFF">-n</font> <font color="#00AFFF">10</font>
<font color="#EC00FF"><b>    automesh 0.2.7</b></font>
     <font color="#58C7E2"><b>Reading</b></font> spheres_resolution_3.npy
        <font color="#54E484"><b>Done</b></font> 463.838µs
     <font color="#58C7E2"><b>Meshing</b></font> target/foo.exo
           <font color="#E0B401"><b>⤷ Removed voxels</b></font> 38.844493ms
             <font color="#E0B401"><b>Element-to-node connectivity</b></font> 38.433797ms
             <font color="#E0B401"><b>Nodal coordinates</b></font> 82.679626ms
             <font color="#E0B401"><b>Renumbered nodes</b></font> 9.065267ms
        <font color="#54E484"><b>Done</b></font> 171.121497ms
   <font color="#58C7E2"><b>Smoothing</b></font> target/foo.exo
           <font color="#E0B401"><b>⤷ Node-to-element connectivity</b></font> 62.162447ms 
             <font color="#E0B401"><b>Node-to-node connectivity</b></font> 257.825856ms 
             <font color="#E0B401"><b>Nodal influencers</b></font> 54.363082ms 
             <font color="#E0B401"><b>Smoothing iteration 1</b></font> 42.681891ms 
             <font color="#E0B401"><b>Smoothing iteration 2</b></font> 48.380029ms 
             <font color="#E0B401"><b>Smoothing iteration 3</b></font> 46.085688ms 
             <font color="#E0B401"><b>Smoothing iteration 4</b></font> 51.0272ms 
             <font color="#E0B401"><b>Smoothing iteration 5</b></font> 34.238263ms 
             <font color="#E0B401"><b>Smoothing iteration 6</b></font> 37.369853ms 
             <font color="#E0B401"><b>Smoothing iteration 7</b></font> 34.17091ms 
             <font color="#E0B401"><b>Smoothing iteration 8</b></font> 37.471688ms 
             <font color="#E0B401"><b>Smoothing iteration 9</b></font> 34.012949ms 
             <font color="#E0B401"><b>Smoothing iteration 10</b></font> 37.26944ms 
        <font color="#54E484"><b>Done</b></font> 788.643428ms
     <font color="#58C7E2"><b>Metrics</b></font> target/foo.npy
           <font color="#E0B401"><b>⤷ Minimum scaled Jacobians</b></font> 212.083563ms
             <font color="#E0B401"><b>Writing metrics to file</b></font> 5.911884ms
        <font color="#54E484"><b>Done</b></font> 265.199271ms
     <font color="#58C7E2"><b>Writing</b></font> target/foo.exo
             <font color="#E0B401"><b>Element-to-node connectivity</b></font> 56.269322ms
           <font color="#E0B401"><b>⤷ Nodal coordinates</b></font> 18.653303ms
        <font color="#54E484"><b>Done</b></font> 166.589199ms
</pre>

## After

<pre>test block_16::calculate_laplacian                 ... <font color="#58C7E2">bench</font>:      19,695.47 ns/iter (+/- 819.38)
test block_16::smooth_laplace                      ... <font color="#58C7E2">bench</font>:      18,732.97 ns/iter (+/- 187.57)
test block_32::calculate_laplacian                 ... <font color="#58C7E2">bench</font>:     165,021.38 ns/iter (+/- 5,479.53)
test block_32::smooth_laplace                      ... <font color="#58C7E2">bench</font>:     141,955.95 ns/iter (+/- 2,108.98)
test block_8::calculate_laplacian                  ... <font color="#58C7E2">bench</font>:       2,369.27 ns/iter (+/- 9.30)
test block_8::smooth_laplace                       ... <font color="#58C7E2">bench</font>:       2,745.21 ns/iter (+/- 8.62)
</pre>

<pre>test block_16::write_metrics_csv                   ... <font color="#58C7E2">bench</font>:     548,785.56 ns/iter (+/- 86,619.74)
test block_16::write_metrics_npy                   ... <font color="#58C7E2">bench</font>:     334,389.40 ns/iter (+/- 66,840.07)
test block_32::write_metrics_csv                   ... <font color="#58C7E2">bench</font>:   3,977,710.70 ns/iter (+/- 40,259.97)
test block_32::write_metrics_npy                   ... <font color="#58C7E2">bench</font>:   2,158,881.50 ns/iter (+/- 45,996.59)
test block_8::write_metrics_csv                    ... <font color="#58C7E2">bench</font>:     118,256.11 ns/iter (+/- 26,479.22)
test block_8::write_metrics_npy                    ... <font color="#58C7E2">bench</font>:      86,556.96 ns/iter (+/- 12,886.03)</pre>

<pre><font color="#005FD7">cargo</font> <font color="#00AFFF">run</font> <font color="#00AFFF">-qr</font> <font color="#00AFFF">-F</font> <font color="#00AFFF">profile</font> <font color="#00AFFF">--</font> <font color="#00AFFF">mesh</font> <font color="#00AFFF">-i</font> <font color="#00AFFF"><u style="text-decoration-style:single">spheres_resolution_3.npy</u></font> <font color="#00AFFF">-o</font> <font color="#00AFFF">target/foo.exo</font> <font color="#00AFFF">--metrics</font> <font color="#00AFFF">target/foo.npy</font> <font color="#00AFFF">smooth</font> <font color="#00AFFF">-n</font> <font color="#00AFFF">10</font>
<font color="#EC00FF"><b>    automesh 0.2.7</b></font>
     <font color="#58C7E2"><b>Reading</b></font> spheres_resolution_3.npy
        <font color="#54E484"><b>Done</b></font> 503.108µs
     <font color="#58C7E2"><b>Meshing</b></font> target/foo.exo
           <font color="#E0B401"><b>⤷ Removed voxels</b></font> 40.121834ms
             <font color="#E0B401"><b>Element-to-node connectivity</b></font> 38.272291ms
             <font color="#E0B401"><b>Nodal coordinates</b></font> 17.801103ms
             <font color="#E0B401"><b>Renumbered nodes</b></font> 9.943699ms
        <font color="#54E484"><b>Done</b></font> 108.547071ms
   <font color="#58C7E2"><b>Smoothing</b></font> target/foo.exo
           <font color="#E0B401"><b>⤷ Node-to-element connectivity</b></font> 61.433314ms 
             <font color="#E0B401"><b>Node-to-node connectivity</b></font> 237.271061ms 
             <font color="#E0B401"><b>Nodal influencers</b></font> 55.432624ms 
             <font color="#E0B401"><b>Smoothing iteration 1</b></font> 13.98701ms 
             <font color="#E0B401"><b>Smoothing iteration 2</b></font> 15.018204ms 
             <font color="#E0B401"><b>Smoothing iteration 3</b></font> 14.888706ms 
             <font color="#E0B401"><b>Smoothing iteration 4</b></font> 13.716988ms 
             <font color="#E0B401"><b>Smoothing iteration 5</b></font> 7.602194ms 
             <font color="#E0B401"><b>Smoothing iteration 6</b></font> 7.368757ms 
             <font color="#E0B401"><b>Smoothing iteration 7</b></font> 7.451234ms 
             <font color="#E0B401"><b>Smoothing iteration 8</b></font> 7.449271ms 
             <font color="#E0B401"><b>Smoothing iteration 9</b></font> 7.270428ms 
             <font color="#E0B401"><b>Smoothing iteration 10</b></font> 7.269586ms 
        <font color="#54E484"><b>Done</b></font> 458.649532ms
     <font color="#58C7E2"><b>Metrics</b></font> target/foo.npy
           <font color="#E0B401"><b>⤷ Minimum scaled Jacobians</b></font> 54.203469ms
             <font color="#E0B401"><b>Writing metrics to file</b></font> 3.635139ms
        <font color="#54E484"><b>Done</b></font> 57.864387ms
     <font color="#58C7E2"><b>Writing</b></font> target/foo.exo
             <font color="#E0B401"><b>Element-to-node connectivity</b></font> 55.83564ms
           <font color="#E0B401"><b>⤷ Nodal coordinates</b></font> 14.105611ms
        <font color="#54E484"><b>Done</b></font> 141.615558ms
</pre>